requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
numpy==1.24.3
psycopg2==2.9.6
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
yarg==0.1.9
